requirements.txt
flask
gunicorn
apscheduler
requests
python-dotenv
psycopg2-binary
cloudinary
bcrypt
stripe
qrcode
Pillow
twilio
anthropic
pyjwt
flask-cors
